+++ lib/karafka/routing/consumer_group.rb
@@ -10,6 +10,12 @@ module Karafka
     class ConsumerGroup
       attr_reader :id, :topics, :name
 
+      # This is a "virtual" attribute that is not building subscription groups.
+      # It allows us to store the "current" subscription group defined in the routing
+      # This subscription group id is then injected into topics, so we can compute the subscription
+      # groups
+      attr_accessor :current_subscription_group_name
+
       # @param name [String, Symbol] raw name of this consumer group. Raw means, that it does not
       #   yet have an application client_id namespace, this will be added here by default.
       #   We add it to make a multi-system development easier for people that don't use
@@ -32,7 +38,27 @@ module Karafka
       def topic=(name, &block)
         topic = Topic.new(name, self)
         @topics << Proxy.new(topic, &block).target
-        @topics.last
+        built_topic = @topics.last
+        # We overwrite it conditionally in case it was not set by the user inline in the topic
+        # block definition
+        built_topic.subscription_group ||= current_subscription_group_name
+        built_topic
+      end
+
+      # Assigns the current subscription group id based on the defined one and allows for further
+      # topic definition
+      # @param name [String, Symbol]
+      # @param block [Proc] block that may include topics definitions
+      def subscription_group=(name, &block)
+        # We cast it here, so the routing supports symbol based but that's anyhow later on
+        # validated as a string
+        self.current_subscription_group_name = name.to_s
+
+        Proxy.new(self, &block)
+
+        # We need to reset the current subscription group after it is used, so it won't leak
+        # outside to other topics that would be defined without a defined subscription group
+        self.current_subscription_group_name = nil
       end
 
       # @return [Array<Routing::SubscriptionGroup>] all the subscription groups build based on