+++ spec/integrations/instrumentation/statistics_on_lag_with_many_sub_group.rb
@@ -42,14 +42,16 @@ Karafka::App.monitor.subscribe('statistics.emitted') do |event|
 end
 
 draw_routes do
-  topic DT.topics[0] do
-    consumer Consumer
-    subscription_group '1'
+  subscription_group '1' do
+    topic DT.topics[0] do
+      consumer Consumer
+    end
   end
 
-  topic DT.topics[1] do
-    consumer Consumer
-    subscription_group '2'
+  subscription_group '2' do
+    topic DT.topics[1] do
+      consumer Consumer
+    end
   end
 end